Move common types into separate crate #6
Since we are accessing logs from both the TUI and the embedded software,
we need to share the types between the two. This commit moves the types
into a separate crate.
