+++ ollyfc/Cargo.toml
@@ -22,6 +22,8 @@ mpu6050-dmp = { git= "https://github.com/barafael/mpu6050-dmp-rs.git"}
 usbd-serial = {version= "0.2.0"}
 usb-device = {version= "0.3.1"}
 
+ollyfc_common = { path = "../ollyfc_common", features=["defmt"]}
+
 [profile.dev]
 codegen-units = 1
 debug = 2
