--- ollyfc/src/lib.rs
@@ -1,3 +1,0 @@
-#![no_std]
-
-pub mod w25q;
