+++ ollyfc_common/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ollyfc_common"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+defmt = { version = "0.3.5", optional = true }
+
+[features]
+default = ["std"]
+use_defmt = ["defmt"]
+std = []
